FE: style & add Icon to btnLogout

+++ frontend/src/js/modules/App.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from "react";
+import React, {} from "react";
 import {
     BrowserRouter as Router,
     Switch,
@@ -47,11 +47,14 @@ export const App = () => {
 
     const btnLogout = (e) => {
         e.preventDefault()
-        deleteCookie('token')
-        document.location.href = '/login'
+        if (confirm("Voulez vous vraiment vous déconnectez ?")) {
+            deleteCookie('token')
+            document.location.href = '/login'
+        }
     }
 
     return <>
+
         <Router>
             <nav className="navbar navbar-expand-lg navbar-light bg-light">
                 <div className="container-fluid">
@@ -68,7 +71,18 @@ export const App = () => {
                             {
                                 userIsLoggedIn ? <>
                                     <LinkMenu to="/messagerie" label={"Messagerie"}></LinkMenu>
-                                    <button className={"btn"} onClick={btnLogout} id={"linkLogout"}>Déconnexion</button>
+                                    <button className={"btn"} id={"btnAccount"} data-bs-toggle="modal"
+                                            data-bs-target="#modalAccount">Profil
+                                    </button>
+                                    <button className={"btn"} onClick={btnLogout} id={"linkLogout"}>
+                                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16"
+                                             fill="currentColor" className="bi bi-box-arrow-right" viewBox="0 0 16 16">
+                                            <path fillRule="evenodd"
+                                                  d="M10 12.5a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-9a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v2a.5.5 0 0 0 1 0v-2A1.5 1.5 0 0 0 9.5 2h-8A1.5 1.5 0 0 0 0 3.5v9A1.5 1.5 0 0 0 1.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-2a.5.5 0 0 0-1 0v2z"/>
+                                            <path fillRule="evenodd"
+                                                  d="M15.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 0 0-.708.708L14.293 7.5H5.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z"/>
+                                        </svg>
+                                    </button>
                                 </> : <>
                                     <LinkMenu to="/login"
                                               label={"Connexion"}>Connexion</LinkMenu>
@@ -80,7 +94,6 @@ export const App = () => {
                     </div>
                 </div>
             </nav>
-
             <Switch>
                 {
                     userIsLoggedIn ? <>
